+++ zrml/prediction-markets/src/migrations.rs
@@ -16,31 +16,19 @@
 // along with Zeitgeist. If not, see <https://www.gnu.org/licenses/>.
 
 use crate::{Config, Pallet};
-#[cfg(feature = "try-runtime")]
-use alloc::string::ToString;
 use alloc::vec::Vec;
 use frame_support::{
     dispatch::Weight,
     log,
-    migration::{put_storage_value, storage_iter},
     pallet_prelude::PhantomData,
     traits::{Get, OnRuntimeUpgrade, StorageVersion},
 };
-#[cfg(feature = "try-runtime")]
-use scale_info::prelude::format;
 use sp_runtime::traits::{One, Saturating};
-use zeitgeist_primitives::types::{AuthorityReport, MarketDisputeMechanism, OutcomeReport};
-use zrml_authorized::Pallet as AuthorizedPallet;
+use zeitgeist_primitives::types::MarketDisputeMechanism;
 use zrml_market_commons::MarketCommonsPalletApi;
 
 const PREDICTION_MARKETS_REQUIRED_STORAGE_VERSION: u16 = 6;
 const PREDICTION_MARKETS_NEXT_STORAGE_VERSION: u16 = 7;
-
-const AUTHORIZED: &[u8] = b"Authorized";
-const AUTHORIZED_OUTCOME_REPORTS: &[u8] = b"AuthorizedOutcomeReports";
-
-const AUTHORIZED_REQUIRED_STORAGE_VERSION: u16 = 2;
-const AUTHORIZED_NEXT_STORAGE_VERSION: u16 = 3;
 
 pub struct UpdateMarketIdsPerDisputeBlock<T>(PhantomData<T>);
 
@@ -316,187 +304,6 @@ mod tests_auto_resolution {
 
     fn set_up_chain() {
         StorageVersion::new(PREDICTION_MARKETS_REQUIRED_STORAGE_VERSION).put::<Pallet<Runtime>>();
-    }
-}
-
-pub struct AddFieldToAuthorityReport<T>(PhantomData<T>);
-
-// Add resolve_at block number value field to `AuthorizedOutcomeReports` map.
-impl<T: Config + zrml_market_commons::Config + zrml_authorized::Config> OnRuntimeUpgrade
-    for AddFieldToAuthorityReport<T>
-{
-    fn on_runtime_upgrade() -> Weight
-    where
-        T: Config,
-    {
-        let mut total_weight = T::DbWeight::get().reads(1);
-        let authorized_version = StorageVersion::get::<AuthorizedPallet<T>>();
-        if authorized_version != AUTHORIZED_REQUIRED_STORAGE_VERSION {
-            log::info!(
-                "AddFieldToAuthorityReport: authorized version is {:?}, require {:?};",
-                authorized_version,
-                AUTHORIZED_REQUIRED_STORAGE_VERSION,
-            );
-            return total_weight;
-        }
-        log::info!("AddFieldToAuthorityReport: Starting...");
-
-        let mut new_storage_map = Vec::new();
-        let now = <frame_system::Pallet<T>>::block_number();
-        for (key, old_value) in
-            storage_iter::<Option<OutcomeReport>>(AUTHORIZED, AUTHORIZED_OUTCOME_REPORTS)
-        {
-            total_weight = total_weight.saturating_add(T::DbWeight::get().reads(1));
-
-            if let Some(outcome) = old_value {
-                let resolve_at: T::BlockNumber =
-                    now.saturating_add(<T as zrml_authorized::Config>::ReportPeriod::get());
-                let new_value = AuthorityReport { resolve_at, outcome };
-                new_storage_map.push((key, new_value));
-            }
-        }
-
-        for (key, new_value) in new_storage_map {
-            put_storage_value::<Option<AuthorityReport<T::BlockNumber>>>(
-                AUTHORIZED,
-                AUTHORIZED_OUTCOME_REPORTS,
-                &key,
-                Some(new_value),
-            );
-            total_weight = total_weight.saturating_add(T::DbWeight::get().writes(1));
-        }
-
-        StorageVersion::new(AUTHORIZED_NEXT_STORAGE_VERSION).put::<AuthorizedPallet<T>>();
-        total_weight = total_weight.saturating_add(T::DbWeight::get().writes(1));
-        log::info!("AddFieldToAuthorityReport: Done!");
-        total_weight
-    }
-
-    #[cfg(feature = "try-runtime")]
-    fn pre_upgrade() -> Result<(), &'static str> {
-        use frame_support::traits::OnRuntimeUpgradeHelpersExt;
-        let counter_key = "counter_key".to_string();
-        Self::set_temp_storage(0_u32, &counter_key);
-        for (key, value) in
-            storage_iter::<Option<OutcomeReport>>(AUTHORIZED, AUTHORIZED_OUTCOME_REPORTS)
-        {
-            Self::set_temp_storage(value, &format!("{:?}", key.as_slice()));
-
-            let counter: u32 =
-                Self::get_temp_storage(&counter_key).expect("counter key storage not found");
-            Self::set_temp_storage(counter + 1_u32, &counter_key);
-        }
-        Ok(())
-    }
-
-    #[cfg(feature = "try-runtime")]
-    fn post_upgrade() -> Result<(), &'static str> {
-        use frame_support::traits::OnRuntimeUpgradeHelpersExt;
-        let mut markets_count = 0_u32;
-        let old_counter_key = "counter_key".to_string();
-        for (key, new_value) in storage_iter::<Option<AuthorityReport<T::BlockNumber>>>(
-            AUTHORIZED,
-            AUTHORIZED_OUTCOME_REPORTS,
-        ) {
-            let key_str = format!("{:?}", key.as_slice());
-            if let Some(AuthorityReport { resolve_at: _, outcome }) = new_value {
-                let old_value: Option<OutcomeReport> = Self::get_temp_storage(&key_str)
-                    .unwrap_or_else(|| panic!("old value not found for market id {:?}", key_str));
-
-                assert_eq!(old_value.unwrap(), outcome);
-
-                markets_count += 1_u32;
-            } else {
-                panic!(
-                    "For market id {:?} storage iter should only find present (Option::Some) \
-                     values",
-                    key_str
-                );
-            }
-        }
-        let old_markets_count: u32 =
-            Self::get_temp_storage(&old_counter_key).expect("old counter key storage not found");
-        assert_eq!(markets_count, old_markets_count);
-        Ok(())
-    }
-}
-
-#[cfg(test)]
-mod tests_authorized {
-    use super::*;
-    use crate::mock::{ExtBuilder, Runtime};
-    use frame_support::Twox64Concat;
-    use zeitgeist_primitives::types::{MarketId, OutcomeReport};
-
-    #[test]
-    fn on_runtime_upgrade_increments_the_storage_versions() {
-        ExtBuilder::default().build().execute_with(|| {
-            set_up_chain();
-            AddFieldToAuthorityReport::<Runtime>::on_runtime_upgrade();
-            let authorized_version = StorageVersion::get::<AuthorizedPallet<Runtime>>();
-            assert_eq!(authorized_version, AUTHORIZED_NEXT_STORAGE_VERSION);
-        });
-    }
-
-    #[test]
-    fn on_runtime_sets_new_struct_with_resolve_at() {
-        ExtBuilder::default().build().execute_with(|| {
-            set_up_chain();
-
-            <frame_system::Pallet<Runtime>>::set_block_number(10_000);
-
-            let hash = crate::migrations::utility::key_to_hash::<Twox64Concat, MarketId>(0);
-            let outcome = OutcomeReport::Categorical(42u16);
-            put_storage_value::<Option<OutcomeReport>>(
-                AUTHORIZED,
-                AUTHORIZED_OUTCOME_REPORTS,
-                &hash,
-                Some(outcome.clone()),
-            );
-
-            AddFieldToAuthorityReport::<Runtime>::on_runtime_upgrade();
-
-            let now = <frame_system::Pallet<Runtime>>::block_number();
-            let resolve_at: <Runtime as frame_system::Config>::BlockNumber =
-                now.saturating_add(<Runtime as zrml_authorized::Config>::ReportPeriod::get());
-            let expected = Some(AuthorityReport { resolve_at, outcome });
-
-            let actual = frame_support::migration::get_storage_value::<
-                Option<AuthorityReport<<Runtime as frame_system::Config>::BlockNumber>>,
-            >(AUTHORIZED, AUTHORIZED_OUTCOME_REPORTS, &hash)
-            .unwrap();
-            assert_eq!(expected, actual);
-        });
-    }
-
-    #[test]
-    fn on_runtime_is_noop_if_versions_are_not_correct() {
-        ExtBuilder::default().build().execute_with(|| {
-            // storage migration already executed (storage version is incremented already)
-            StorageVersion::new(AUTHORIZED_NEXT_STORAGE_VERSION).put::<AuthorizedPallet<Runtime>>();
-
-            let hash = crate::migrations::utility::key_to_hash::<Twox64Concat, MarketId>(0);
-            let outcome = OutcomeReport::Categorical(42u16);
-            let now = <frame_system::Pallet<Runtime>>::block_number();
-            let resolve_at: <Runtime as frame_system::Config>::BlockNumber =
-                now.saturating_add(<Runtime as zrml_authorized::Config>::ReportPeriod::get());
-            let report = AuthorityReport { resolve_at, outcome };
-            put_storage_value::<
-                Option<AuthorityReport<<Runtime as frame_system::Config>::BlockNumber>>,
-            >(AUTHORIZED, AUTHORIZED_OUTCOME_REPORTS, &hash, Some(report.clone()));
-
-            AddFieldToAuthorityReport::<Runtime>::on_runtime_upgrade();
-
-            let actual = frame_support::migration::get_storage_value::<
-                Option<AuthorityReport<<Runtime as frame_system::Config>::BlockNumber>>,
-            >(AUTHORIZED, AUTHORIZED_OUTCOME_REPORTS, &hash)
-            .unwrap();
-            assert_eq!(Some(report), actual);
-        });
-    }
-
-    fn set_up_chain() {
-        StorageVersion::new(AUTHORIZED_REQUIRED_STORAGE_VERSION).put::<AuthorizedPallet<Runtime>>();
     }
 }
